Updated tests

+++ test.js
@@ -6,17 +6,18 @@ describe('test page', function() {
    it('should look like test page', function () {
        browser.url('http://localhost:8080');
        if (fs.existsSync('baselines/myImage.png')) {
+           if (fs.existsSync('shots/myImage.png')) {
             fs.unlink('shots/myImage.png');  
-            browser.saveScreenshot('shots/myImage.png');
+           }
+            browser.saveElementScreenshot('shots/myImage.png', 'p');
             var baseline = fs.readFileSync('baselines/myImage.png');
             var shot = fs.readFileSync('shots/myImage.png');
-            resemble(baseline).compareTo(shot).onComplete(function(data){
+            var result = resemble(baseline).compareTo(shot).onComplete(function(data){
                 expect(data.misMatchPercentage).is.lessThan('0.05');
             });
-
         }
         else {
-            browser.saveScreenshot('./baselines/myImage.png');
+            browser.saveElementScreenshot('./baselines/myImage.png', 'p');
         }
        
 });
